Handle the case when input string contains regexp syntaxis, e.g. [ ]
`mqr-perform-replace' is always called with arg `regexp-flag' non-nil.
This causes problems if the user want to match plain strings, that is, no
regexps, if some includes special regex symbols.

For instance, if the user wants to match the string '[foo',
the command will fail because that is not a valid regexp, and
`mqr-perform-replace' only performs regexp replacement.

One way to fix this issue is add a new variable specifying whether if we are
performing a 'plain strings' match or a regexps one.  Then, `mqr--replacement'
will quote the strings in the case of 'plain strings' match.

In the example above, the effect is to match with '\\[foo' instead of '[foo'.

* mqr.el (mqr--regexp-replace): New variable.

(mqr-replace-regexp, mqr-query-replace-regexp):
Bind the new var to a non-nil value.

(mqr--replacement): Rename argument to matched-string.
Use `regexp-quote' when `mqr--regexp-replace' is nil, that is,
when the user have input plain strings; that's the case
for commands `mqr-query-replace' or `mqr-replace'.

* mqr-tests.el: Add test for this bug.